+++ src/exists/index.ts
@@ -1,0 +1,2 @@
+export { assertExists } from "./assert-exists";
+export { exists } from "./exists";
