+++ src/main.ts
@@ -26,9 +26,20 @@ export async function main(): Promise<boolean> {
 	const config = await importConfig(logger, configPath);
 	const tsFiles = getAllTsFiles(logger, root);
 
-	const { scoped, errorsRef, reports, hasOnlyScopes } = check(
+	const onlyScopes = config.scopes.filter((scope) => scope.only);
+	const hasOnlyScopes = 0 < onlyScopes.length;
+	const scopeDeclarations = hasOnlyScopes ? onlyScopes : config.scopes;
+
+	if (hasOnlyScopes) {
+		console.info(
+			`Found ${onlyScopes.length} scope(s) with "only: true", processing only these scopes`,
+		);
+		console.info(""); // blank
+	}
+
+	const { scoped, errorsRef, reports } = check(
 		logger,
-		config,
+		scopeDeclarations,
 		tsFiles,
 		root,
 	);
@@ -39,17 +50,17 @@ export async function main(): Promise<boolean> {
 		.flatMap((v) => v.result)
 		.filter((v) => !v.isAllowed).length;
 
-	outputFromTree(
-		buildTree(
-			errorsRef,
-			reports,
-			tsFiles,
-			scoped,
-			needsReportUnscoped,
-			duration,
-			restrictedImports,
-		),
+	const tree = buildTree(
+		errorsRef,
+		reports,
+		tsFiles,
+		scoped,
+		needsReportUnscoped,
+		duration,
+		restrictedImports,
 	);
+
+	outputFromTree(tree);
 
 	if (hasOnlyScopes) {
 		console.info(`Failed due to "only: true" flag`);
